refactor: split value.rs and customhash/lib.rs
